+++ src/app/app.routes.ts
@@ -75,10 +75,35 @@ export const routes: Routes = [
   {
     path: 'crew-area',
     loadComponent: () => {
-      return import('./components/crew-area/crew-area.component').then(
-        (m) => m.CrewAreaComponent,
+      return import('./components/crew-area/crew-container.component').then(
+        (m) => m.CrewContainerComponent,
       );
     },
     canActivate: [AuthGuard],
+    children: [
+      {
+        path: '',
+        redirectTo: 'manage-tickets',
+        pathMatch: 'full',
+      },
+      {
+        path: 'manage-tickets',
+        loadComponent: () => {
+          return import(
+            './components/crew-area/manager/manager.component'
+          ).then((m) => m.ManagerComponent);
+        },
+        canActivate: [AuthGuard],
+      },
+      {
+        path: 'scheduler',
+        loadComponent: () => {
+          return import(
+            './components/crew-area/scheduler/scheduler.component'
+          ).then((m) => m.SchedulerComponent);
+        },
+        canActivate: [AuthGuard],
+      },
+    ],
   },
 ];